# api/api.yml
openapi: 3.0.2
info:
  version: 1.0.0
  title: Issue Tracker
  description: "This is the Open API documentation for the REST API of the issue tracker developed for the ASW Lab class. All operations require authentication."
servers:
  - url: 'https://issuetracker-ff8u.onrender.com/'
  - url: 'http://localhost:8000'
paths:
  /issues/:
    get:
      tags:
        - issues
      summary: List and filter issues
      operationId: listIssues
      parameters:
        - name: search
          in: query
          description: Search text for subject or description
          schema:
            type: string
        - name: order_by
          in: query
          description: Field to order by (e.g., priority is asc, -priority is desc)
          schema:
            type: string
            enum: ['issue_type', '-issue_type', 'issue_severity', '-issue_severity', 'priority', '-priority', 'subject', '-subject', 'status', '-status', 'assignee', '-assignee', 'modified_at', '-modified_at', 'deadline', '-deadline', 'created_at', '-created_at']
        - name: issue_type
          in: query
          description: "Filter by issue type"
          explode: true
          schema:
            type: array
            items:
              type: string
              enum: [Bug, Question, Enhancement]
        - name: issue_severity
          in: query
          description: "Filter by issue severity"
          explode: true
          schema:
            type: array
            items:
              type: string
              enum: [Wishlist, Minor, Normal, Important, Critical]
        - name: status
          in: query
          description: "Filter by status"
          explode: true
          schema:
            type: array
            items:
              type: string
              enum: [New, In Progress, Ready for test, Needs Info, Rejected, Postponed, Closed]

        - name: assigned_to
          in: query
          description: "User ID or 'unassigned'"
          schema:
            type: string

      responses:
        '200':
          description: Issues retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  issues:
                    type: array
                    items:
                      $ref: '#/components/schemas/Issue'
                  current_order:
                    type: string
                  unassigned_count:
                    type: integer
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
      security:
        - ApiKeyAuth: []
    post:
      tags:
        - issues
      summary: Create a new issue
      operationId: createIssue
      requestBody:
        description: Provide the details for the new issue
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                subject:
                  type: string
                  example: "Bug in login"
                description:
                  type: string
                  example: "The login button is not working"
                assignee:
                  type: integer
                  example: "1"
                deadline:
                  type: string
                  format: date
                  example: '2026-05-08'
                issue_type:
                  type: string
                  enum: [Bug, Question, Enhancement]
                  example: "Bug"
                priority:
                  type: string
                  enum: [Low, Normal, High]
                  example: "Normal"
                issue_severity:
                  type: string
                  enum: [Wishlist, Minor, Normal, Important, Critical]
                  example: "Normal"
                status:
                  type: string
                  enum: [New, In Progress, Ready for test, Needs Info, Rejected, Postponed, Closed]
                  example: "New"
                files:
                  type: string
                  format: binary
              required:
                - subject
        required: true
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          description: Bad Request (Subject is required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
      security:
        - ApiKeyAuth: []
...
  /issue/{issueId}/:
    get:
      tags:
        - issues
      summary: Get issue details
      operationId: getIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Issue details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
      security:
        - ApiKeyAuth: []
    put:
      tags:
        - issues
      summary: Update an issue
      description: |
        Updates an issue. All fields are optional; only provided fields are modified.

        **Creator-only fields** (`subject`, `description`, `status`, `issue_type`, `issue_severity`, `priority`, `tags`, `deadline`): only the issue creator can send these. Returns 403 if a non-creator includes any of them.

        **Any authenticated user** can modify `assignee` and `watchers`.

        **Attachments**: any authenticated user can add or delete attachments via the dedicated endpoints (`POST /issue/{issueId}/attachments` and `DELETE /attachments/{attachmentId}`).
      operationId: updateIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Fields to update (all optional; only provided fields are modified)
        required: true
        content:
          application/json:
            example:
              subject: "Updated bug title"
              description: "More details about the bug"
              status: 2
              issue_type: 1
              issue_severity: 3
              priority: 2
              tags: [1]
              deadline: "2026-06-01"
              assignee: 2
              watchers: [1, 2]
            schema:
              type: object
              properties:
                subject:
                  type: string
                  description: "[Creator only] New title for the issue"
                  example: "Updated bug title"
                description:
                  type: string
                  description: "[Creator only] New description"
                  example: "More details about the bug"
                status:
                  type: integer
                  description: "[Creator only] ID of the status"
                  example: 2
                issue_type:
                  type: integer
                  description: "[Creator only] ID of the type"
                  example: 1
                issue_severity:
                  type: integer
                  description: "[Creator only] ID of the severity"
                  example: 3
                priority:
                  type: integer
                  description: "[Creator only] ID of the priority"
                  example: 2
                tags:
                  type: array
                  description: "[Creator only] List of tag IDs (replaces current tags)"
                  items:
                    type: integer
                  example: [1, 3]
                deadline:
                  type: string
                  format: date
                  nullable: true
                  description: "[Creator only] Deadline date, or null to remove it"
                  example: "2026-06-01"
                assignee:
                  type: integer
                  nullable: true
                  description: "ID of the user to assign, or null to unassign"
                  example: 5
                watchers:
                  type: array
                  description: "List of user IDs that will watch the issue (replaces current watchers)"
                  items:
                    type: integer
                  example: [2, 7]

      responses:
        '200':
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '400':
          description: Bad request (e.g. empty subject, unknown ID, or invalid list)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        '403':
          description: Forbidden (non-creator attempted to edit a creator-only field)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
      security:
        - ApiKeyAuth: []

    delete:
      tags:
        - issues
      summary: Delete an issue
      operationId: deleteIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Issue deleted
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        '403':
          description: Invalid user (Only creator can delete)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
...
    Issue:
      type: object
      properties:
        id:
          type: integer
          example: 1
        subject:
          type: string
          example: "Bug in login"
        status:
          type: string
          example: "Open"
        priority:
          type: string
          example: "High"
        severity:
          type: string
          example: "Normal"
        assignee:
          type: string
          example: "Andreu-Caro"
        deadline:
          $ref: '#/components/schemas/Deadline'
        created_at:
          type: string
          format: date-time
          example: "2026-05-04T21:00:00Z"
    IssueDetail:
      allOf:
        - $ref: '#/components/schemas/Issue'
        - type: object
          properties:
            description:
              type: string
              example: "The login button is not working on Chrome"
            current_watchers_count:
              type: integer
              example: 5
            watchers_list:
              type: array
              items:
                type: string
              example: [ "Christian-Barone", "Andreu-Caro" ]
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
            tags:
              type: array
              items:
                type: string
                example: "frontend"
...
    Deadline:
      type: object
      properties:
        deadline:
          type: string
          format: date
          example: "2026-05-08"
...
    Attachment:
      type: object
      properties:
        attachment_id:
          type: integer
          format: int64
          example: 2
        issue_id:
          type: integer
          format: int64
          example: 6
        creator_id:
          type: integer
          format: int64
          example: 8
        url:
          type: string
          format: url
          example: 'https://storage.googleapis.com/issue-tracker2026-media/attachments/barcelona.webp'
        name:
          type: string
          example: 'Barcelona'
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        body:
          type: string
          example: "This is a comment body"
        author:
          type: string
          example: "john_doe"
        created_at:
          type: string
          format: date-time
          example: "2026-05-04T21:00:00Z"
        issue_id:
          type: integer
          format: int64
          example: 6
...
    ErrorResult:
      type: object
      properties:
        message:
          type: string
          example: 'Content is too long (maximum is 280 characters)'
      required:
        - message
  examples:
...
    error400:
      value:
        message: "Body is required"
    error401:
      value:
        message: "The API key you provided does not belong to any users"
    error403:
      value:
        message: "The provided API key does not authorize this action"
    error404:
      value:
        message: "There is no resource with 'id'=8"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header